Various minor inconsistencies (mostly to do with subtle variations in types)
git-svn-id: file:///home/hardaker/lib/sf-bkups/net-snmp-convert-svnrepo/trunk@6703 06827809-a52a-0410-b366-d66718629ded

+++ include/net-snmp/library/snmpUDPDomain.h
@@ -32,7 +32,7 @@ void		netsnmp_udp_parse_security	(const char *token, char *param);
 
 int		netsnmp_udp_getSecName	(void *opaque, int olength,
 					 const char *community,
-					 int community_len, char **secname);
+					 size_t community_len, char **secname);
 
 /*  "Constructor" for transport domain object.  */
